Density models code for NFW, DK, and BMO profiles
Ths structure of this code is likely to change. "Profile1D" (profile.py) will take the place of the "profile" class in the future. NFW has a factor of h that needs to be accounted for and DK profile needs to be fully tested after getting rid fo h dependency on input/outputs